devtools/server/tests/chrome/test_memory.html

<!DOCTYPE HTML>
<html>
<!--
Bug 923275 - Add a memory monitor widget to the developer toolbar
-->
<head>
  <meta charset="utf-8">
  <title>Memory monitoring actor test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="memory-helpers.js" type="application/javascript"></script>
<script>
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  (async function() {
    const { memory, target } = await startServerAndGetSelectedTabMemory();
    const measurement = await memory.measure();
    ok(measurement.total > 0, "total memory is valid");
    ok(measurement.domSize > 0, "domSize is valid");
    ok(measurement.styleSize > 0, "styleSize is valid");
    ok(measurement.jsObjectsSize > 0, "jsObjectsSize is valid");
    ok(measurement.jsStringsSize > 0, "jsStringsSize is valid");
    ok(measurement.jsOtherSize > 0, "jsOtherSize is valid");
    ok(measurement.otherSize > 0, "otherSize is valid");
    ok(measurement.jsMilliseconds, "jsMilliseconds is valid");
    ok(measurement.nonJSMilliseconds, "nonJSMilliseconds is valid");
    await destroyServerAndFinish(target);
  })();
};
</script>
</pre>
</body>
</html>
